leap.cu: fix bug with classification of critical events
Take

1 + 1 -> 2 0

and state
10000 0
0 10000

every specie is part of a critical reaction, in every subvolume,
so tau_ncr = Inf. But tau_cr is also Inf because we skip the
tau computation (and return Inf) if skip = true, and skip is
the combination of "skip because critical reaction" and "skip
because critial diffusion". So every subvolume has both tau_ncr
and tau_cr set to Inf.

Fix this by separating skip into skip_r and skip_d. Even if the
specie is part of a critical reaction, allow it to diffuse if
the diffusion channels are not critical.


Former-commit-id: e628aeaf9819a18e09efff4d3659767616f83482